Move common code from aidants_connect to aidants_connect_common
