+++ aidants_connect_common/static/js/utils.js
