+++ aidants_connect_habilitation/apps.py
@@ -5,5 +5,5 @@ class AidantConnectHabilitationConfig(AppConfig):
     name = "aidants_connect_habilitation"
 
     def ready(self):
-        import aidants_connect.common.lookups  # noqa
+        import aidants_connect_common.lookups  # noqa
         import aidants_connect_habilitation.signals  # noqa
