+++ aidants_connect_web/apps.py
@@ -5,5 +5,5 @@ class AidantConnectWebConfig(AppConfig):
     name = "aidants_connect_web"
 
     def ready(self):
-        import aidants_connect.common.lookups  # noqa
+        import aidants_connect_common.lookups  # noqa
         import aidants_connect_web.signals  # noqa
