+++ aidants_connect_web/forms.py
@@ -12,7 +12,7 @@ from magicauth.forms import EmailForm as MagicAuthEmailForm
 from phonenumber_field.formfields import PhoneNumberField
 from phonenumber_field.widgets import PhoneNumberInternationalFallbackWidget
 
-from aidants_connect.common.constants import AuthorizationDurations as ADKW
+from aidants_connect_common.utils.constants import AuthorizationDurations as ADKW
 from aidants_connect_web.models import (
     Aidant,
     CarteTOTP,
