+++ aidants_connect_web/tests/test_functional/test_404.py
@@ -2,7 +2,7 @@ from django.test import tag
 
 from selenium.webdriver.common.by import By
 
-from aidants_connect.common.tests.testcases import FunctionalTestCase
+from aidants_connect_common.tests.testcases import FunctionalTestCase
 
 
 @tag("functional")
