+++ aidants_connect_web/tests/test_functional/test_a11y.py
@@ -3,7 +3,7 @@ from django.test import tag
 from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.common.by import By
 
-from aidants_connect.common.tests.testcases import FunctionalTestCase
+from aidants_connect_common.tests.testcases import FunctionalTestCase
 
 
 @tag("functional", "a11y")  # nb: a11y stands for "accessibility"
