+++ aidants_connect_web/tests/test_functional/test_admin_transfert_mandat.py
@@ -10,7 +10,7 @@ from selenium.webdriver.support.expected_conditions import url_matches
 from selenium.webdriver.support.select import Select
 from selenium.webdriver.support.wait import WebDriverWait
 
-from aidants_connect.common.tests.testcases import FunctionalTestCase
+from aidants_connect_common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.admin import MandatAdmin
 from aidants_connect_web.models import Mandat
 from aidants_connect_web.tests.factories import AidantFactory, MandatFactory
