+++ aidants_connect_web/tests/test_functional/test_cancel_autorisation.py
@@ -6,7 +6,7 @@ from django.utils import timezone
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.wait import WebDriverWait
 
-from aidants_connect.common.tests.testcases import FunctionalTestCase
+from aidants_connect_common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.models import Journal
 from aidants_connect_web.tests.factories import (
     AidantFactory,
