+++ aidants_connect_web/tests/test_functional/test_filter_in_admin.py
@@ -4,9 +4,9 @@ from django.test import tag
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.wait import WebDriverWait
 
-from aidants_connect.common.tests.testcases import FunctionalTestCase
 from aidants_connect_common.models import Region
 from aidants_connect_common.tests.factories import DepartmentFactory, RegionFactory
+from aidants_connect_common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.tests.factories import (
     AidantFactory,
     HabilitationRequestFactory,
