+++ aidants_connect_web/tests/test_functional/test_import_aidant.py
@@ -10,7 +10,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support.wait import WebDriverWait
 
 import aidants_connect_web
-from aidants_connect.common.tests.testcases import FunctionalTestCase
+from aidants_connect_common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.models import Aidant
 from aidants_connect_web.tests.factories import (
     AidantFactory,
