+++ aidants_connect_web/tests/test_functional/test_login.py
@@ -4,7 +4,7 @@ from django.test import tag
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 
-from aidants_connect.common.tests.testcases import FunctionalTestCase
+from aidants_connect_common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.tests.factories import AidantFactory
