+++ aidants_connect_web/tests/test_functional/test_renew_mandat.py
@@ -5,7 +5,7 @@ from django.utils import timezone
 
 from selenium.webdriver.common.by import By
 
-from aidants_connect.common.tests.testcases import FunctionalTestCase
+from aidants_connect_common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.models import Mandat
 from aidants_connect_web.tests.factories import (
     AidantFactory,
