+++ aidants_connect_web/tests/test_functional/test_usagers.py
@@ -1,6 +1,6 @@
 from selenium.webdriver.common.by import By
 
-from aidants_connect.common.tests.testcases import FunctionalTestCase
+from aidants_connect_common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.tests.factories import (
     AidantFactory,
     ExpiredMandatFactory,
