+++ aidants_connect_web/tests/test_functional/test_view_autorisations.py
@@ -5,7 +5,7 @@ from django.utils import timezone
 
 from selenium.webdriver.common.by import By
 
-from aidants_connect.common.tests.testcases import FunctionalTestCase
+from aidants_connect_common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.tests.factories import (
     AidantFactory,
     AutorisationFactory,
